% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_support_files.R
\name{check_support_files}
\alias{check_support_files}
\title{check_support_files(support_files_path)}
\usage{
check_support_files(support_files_path)
}
\arguments{
\item{support_files_path}{The absolute file path to support files directory.}
}
\value{
Returns a list of files which are not correct.
}
\description{
Checks DataPack support files to ensure they are compatible
with this version of the code.
}
