% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{clusterDistribution}
\alias{clusterDistribution}
\title{clusterDistribution()}
\usage{
clusterDistribution(df)
}
\arguments{
\item{df}{Name of distribution source file.}
}
\value{
Returns a dataframe ready for use in distributing site level targets.
}
\description{
Creates PSNU-level percentages against which Data Pack values can be multiplied for distribution purposes
}
