% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusters}
\alias{clusters}
\title{clusters}
\format{An object of class \code{data.frame} with 51 rows and 12 columns.}
\usage{
clusters
}
\value{
Returns an object with the following columns
}
\description{
Object used dealing with distribution of values from clusters to PSNUs
}
\keyword{datasets}
