% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COP18deMapT}
\alias{COP18deMapT}
\title{COP18deMapT}
\format{An object of class \code{data.frame} with 5275 rows and 23 columns.}
\usage{
COP18deMapT
}
\value{
Returns an object with the following columns
}
\description{
Object used for further processing and matching of data pack codes
}
\keyword{datasets}
