% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPackImporter.R
\docType{package}
\name{datapackimporter}
\alias{datapackimporter}
\alias{datapackimporter-package}
\title{\code{datapackimporter} package}
\description{
Parsing and import functions for the ICPI Data Pack
https://github.com/jennybc/googlesheets/blob/master/R/googlesheets.R
}
