% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackR.R
\name{DataPackR}
\alias{DataPackR}
\title{DataPackR(wb_path,distribution_method, support_files_path}
\usage{
DataPackR(wb_path, distribution_method, support_files_path)
}
\arguments{
\item{wb_path}{Path to the DataPack file to process}

\item{distribution_method}{Should be either 2017 or 2018}

\item{support_file_path}{Path to the support files}
}
\description{
Process DataPack Disagg and Site tools
}
