% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{des}
\alias{des}
\title{des}
\format{An object of class \code{data.frame} with 1683 rows and 2 columns.}
\usage{
des
}
\value{
Returns a two column data frame with codes and combis
}
\description{
List of DataPack codes and DATIM DE.COC combinations
}
\keyword{datasets}
