% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distributeCluster}
\alias{distributeCluster}
\title{distributeCluster(data)}
\usage{
distributeCluster(d)
}
\arguments{
\item{d}{Data object returned from ImportSheets function.}
}
\value{
Returns an object structured for allocation to the import into the PNSU dataset.
}
\description{
Distributes clustered Data Pack data to PSNU level for DATIM import
}
