% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distributeSite}
\alias{distributeSite}
\title{distributeSite()}
\usage{
distributeSite(df, dm)
}
\arguments{
\item{df}{Name of dataframe storing Data Pack data.}

\item{dm}{Distribution Method selection, whether based on FY17, FY18, or not distributed}
}
\value{
Returns a dataframe structured for DATIM import.
}
\description{
Distributes Data Pack data (both Clustered and non-Clustered) to Site level for DATIM import
}
