% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distrSource}
\alias{distrSource}
\title{distrSource()}
\usage{
distrSource(df, FY = NULL)
}
\arguments{
\item{df}{Name of dataframe storing DATIM extract.}

\item{FY}{FiscalYear for which source distribution should be pulled.}
}
\value{
Returns a dataframe ready for use in preparing source files.
}
\description{
Converts FY17/18 dataElements to FY19 PSNU level elements within DATIM extract
}
