% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_dhis_tibble}
\alias{empty_dhis_tibble}
\title{empty_datim_tibble}
\usage{
empty_dhis_tibble()
}
\value{
An empty tibble with DE, perod, orgunit, category option combo, attribute option combo, value
}
\description{
empty_datim_tibble
}
