% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_site_level_tool.R
\name{export_site_level_tool}
\alias{export_site_level_tool}
\title{export_site_level_tool(d)}
\usage{
export_site_level_tool(d)
}
\arguments{
\item{d}{Object returned from the site level distribution function}
}
\description{
Validates the layout of all relevant sheets in a data pack workbook
}
