% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack-importer.R
\name{GetWorkbookInfo}
\alias{GetWorkbookInfo}
\title{GetWorkbookInfo(wb_path,distribution_method,support_files_path)}
\usage{
GetWorkbookInfo(wb_path, distribution_method = NA, support_files_path = NA)
}
\arguments{
\item{wb_path}{The absolute file path to the workbook.}

\item{distribution_method}{The distribution method to use.}

\item{support_files_path}{Path to the support files directory.}
}
\value{
Returns a list consiting of :
 \itemize{
   \item wb_path: Full path to the disagg tool
   \item timestamp: Timestamp of when this script was run
   \item wb_type: Should either be HTS or NORMAL
   \item ou_name: Name of the operating unit
   \item ou_uid: UID of the operating unit }
}
\description{
Provides information about the workbook
}
