% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack-importer.R
\name{GetWorkbookType}
\alias{GetWorkbookType}
\title{GetWorkbookType(wb_path)}
\usage{
GetWorkbookType(wb_path)
}
\arguments{
\item{wb_path}{The absolute file path to the workbook.}
}
\value{
Returns a character value of either NORMAL or HTS. Exits on anything else.
}
\description{
Determines the workbook type
}
