% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hts_schema}
\alias{hts_schema}
\title{hts_schema}
\format{An object of class \code{list} of length 2.}
\usage{
hts_schema
}
\value{
Returns a list which defines the schema for the HTS Disagg tool

HTS Schemas of the Data Pack Excel sheets
}
\description{
Returns the HTS Schema
}
\keyword{datasets}
