% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hts_site_schema}
\alias{hts_site_schema}
\title{hts_site_schema}
\format{Schema of the HTS site level tool}
\usage{
data(hts_site_schema)
}
\description{
Schema for the HTS Site template
}
\keyword{datasets}
