% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impatt}
\alias{impatt}
\title{impatt}
\format{Returns a three column data frame with code,name and Data Pack code (dp_code)}
\usage{
data(impatt)
}
\description{
Obejct for processing of IMPATT data
}
\keyword{datasets}
