% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_disagg_tool_sheet.R
\name{import_disagg_tool_sheet}
\alias{import_disagg_tool_sheet}
\title{import_disagg_tool_sheet(wb_info,schema)}
\usage{
import_disagg_tool_sheet(wb_info, schema)
}
\arguments{
\item{wb_info}{Workbook info object}

\item{schema}{Schema for a single sheet}
}
\value{
A DHIS2 compatible import data frame
}
\description{
import_disagg_tool_sheet(wb_info,schema)
}
