% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_impatt_sheet.R
\name{import_impatt_sheet}
\alias{import_impatt_sheet}
\title{import_impatt_sheet(wb_info,schema)}
\usage{
import_impatt_sheet(wb_info, schema)
}
\arguments{
\item{wb_info}{Workbook info object}

\item{schema}{Schema for a single sheet}
}
\value{
A DHIS2 compatible import data frame
}
