% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_follow_on_mechs_sheet.R
\name{ImportFollowOnMechs}
\alias{ImportFollowOnMechs}
\title{ImportFollowOnMechs(d)}
\usage{
ImportFollowOnMechs(d)
}
\arguments{
\item{d}{Object from import sheets}
}
\value{
An data object with a follow on mechs sheet object. NULL if blank. 
If this sheet is blank, returns NULL.
}
\description{
Imports the follow on mechs sheet.
}
