% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack-importer.R
\name{ImportSheets}
\alias{ImportSheets}
\title{ImportSheets(wb_path,wb_type)}
\usage{
ImportSheets(wb_path)
}
\arguments{
\item{wb_path}{The absolute file path to the workbook.}
}
\value{
Returns a boolean value TRUE if the the workbook is valid, otherwise false
}
\description{
Imports a single sheet from a workbook.
}
