% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{LoadConfig}
\alias{LoadConfig}
\title{Title}
\usage{
LoadConfig(config_path = NA)
}
\arguments{
\item{config_path}{Absolute file path to the configuration file.}
}
\description{
Loads a standard DATIM configuration file.
}
