% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main_site_schema}
\alias{main_site_schema}
\title{main_site_schema}
\format{An object of class \code{list} of length 2.}
\usage{
main_site_schema
}
\value{
A list of column and row positions
}
\description{
Schema of the Normal site level tool
}
\keyword{datasets}
