% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mechs}
\alias{mechs}
\title{mechs}
\format{An object of class \code{data.frame} with 3304 rows and 2 columns.}
\usage{
mechs
}
\value{
Returns a two column data frame with mechanisms and codes
}
\description{
List of mechanisms and codes
}
\keyword{datasets}
