% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{militaryUnits}
\alias{militaryUnits}
\title{militaryUnits}
\format{An object of class \code{data.frame} with 35 rows and 3 columns.}
\usage{
militaryUnits
}
\value{
Returns a data frame of military units which shold be exlcuded
}
\description{
PSNU UIDs and names
}
\keyword{datasets}
