% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_export_to_datim.R
\name{prepare_export_to_datim}
\alias{prepare_export_to_datim}
\title{prepare_export_to_datim(d)}
\usage{
prepare_export_to_datim(d)
}
\arguments{
\item{d}{Data object from ImportSheets or parse_site_level_tool}
}
\value{
Object with aggregatedHTS data, where applicable. In cases of clustered data, the data will be distributed to PSNU level.
}
\description{
Prepares a PSNU or Site level object for import to DATIM.
}
