% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psnus}
\alias{psnus}
\title{psnus}
\format{An object of class \code{list} of length 37.}
\usage{
psnus
}
\value{
Returns a list of data frames
}
\description{
PSNU UIDs and names
}
\keyword{datasets}
