% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rCOP18deMap}
\alias{rCOP18deMap}
\title{rCOP18deMap}
\format{An object of class \code{data.frame} with 5335 rows and 27 columns.}
\usage{
rCOP18deMap
}
\value{
Returns an object with the following columns
}
\description{
Object used for further processing and matching of data pack codes
}
\keyword{datasets}
