% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rCOP18deMapT}
\alias{rCOP18deMapT}
\title{rCOP18deMapT}
\format{Object used for mapping indicators and disaggs across time, including from FY19 PSNU level to FY19 Site level. Does not translate new indicators/disaggs}
\usage{
data(rCOP18deMapT)
}
\description{
Object for transformation of data element and disagg codes.
}
\keyword{datasets}
