% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{round_trunc}
\alias{round_trunc}
\title{round_trunc(x)}
\usage{
round_trunc(x)
}
\arguments{
\item{x}{Value to be rounded}
}
\value{
Returns a vector of integers with length=length(x)
}
\description{
Rounds values by the rule of half away from zero -- consistent for both positive and negative numbers.
}
