% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{siteDistribution}
\alias{siteDistribution}
\title{siteDistribution()}
\usage{
siteDistribution(df)
}
\arguments{
\item{df}{Name of distribution source file.}
}
\value{
Returns a dataframe ready for use in distributing site level targets.
}
\description{
Creates site-level percentages against which Data Pack values can be multiplied for distribution purposes
}
