% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites_exclude}
\alias{sites_exclude}
\title{sites_exclude}
\format{An object of class \code{character} of length 111.}
\usage{
sites_exclude
}
\value{
Returns a character vector of UIDs
}
\description{
Vector of sites to exlcude
}
\keyword{datasets}
