% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{support_files_md5}
\alias{support_files_md5}
\title{support_files_md5}
\format{Object used for validation of support files}
\usage{
data(support_files_md5)
}
\description{
Checksum of required non-public support files
}
\keyword{datasets}
