% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_impatt_sheet.R
\name{ValidateImpattSheet}
\alias{ValidateImpattSheet}
\title{ValidateImpattSheet(d,wb_info)}
\usage{
ValidateImpattSheet(d, wb_info)
}
\arguments{
\item{d}{A parsed data frame with IMPATT data}

\item{wb_info}{Workbook info for the workbook}
}
\description{
Validates the impatt sheet for completeness.
}
