% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack-importer.R
\name{ValidateSheet}
\alias{ValidateSheet}
\title{ValidateSheet(d,this_sheet)}
\usage{
ValidateSheet(d, this_sheet)
}
\arguments{
\item{d}{Info about the workbook.}

\item{this_sheet}{A particular sheet to validate.}
}
\value{
Returns a boolean value TRUE if the sheet is valid, otherwise, FALSE.
}
\description{
Validates the layout of a single sheet based on its schema definition.
}
