% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack-importer.R
\name{ValidateSheets}
\alias{ValidateSheets}
\title{ValidateSheets(d)}
\usage{
ValidateSheets(d)
}
\arguments{
\item{d}{Parsed data pack object with workbook info}
}
\value{
Returns a boolean named vector of sheets and their validation status.
}
\description{
Validates all of the sheets
}
