% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distributeCluster}
\alias{distributeCluster}
\title{distributeCluster()}
\usage{
distributeCluster(df, distribution_year)
}
\arguments{
\item{df}{Name of dataframe storing Data Pack data.}

\item{distribution_year}{Distribution Method selection, whether based on FY17. Otherwise, FY18.}
}
\value{
Returns a dataframe structured for DATIM import.
}
\description{
Distributes clustered Data Pack data to PSNU level for DATIM import
}
