% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distributeCluster}
\alias{distributeCluster}
\title{distributeCluster(data)}
\usage{
distributeCluster(d, Pcts = NULL)
}
\arguments{
\item{d}{Data object returned from ImportSheets function.}

\item{Pcts}{Percentage object for this data set. Setting this allows one to override the default 
percentage allocation with a custom one.}
}
\value{
Returns an object structured for allocation to the import into the PNSU dataset.
}
\description{
Distributes clustered Data Pack data to PSNU level for DATIM import
}
