% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distributeSite}
\alias{distributeSite}
\title{distributeSite(d)}
\usage{
distributeSite(d, Pcts = NULL)
}
\arguments{
\item{d}{Data object returned from ImportSheets function.}

\item{Pcts}{Percentage object used to distribute from PSNU to site. Setting this allows for the default
distribution method to be overridden.}
}
\value{
Returns an object structured for allocation to the site level tool.
}
\description{
Distributes Data Pack data (both Clustered and non-Clustered) to Site level for DATIM import
}
