% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hts_schema}
\alias{hts_schema}
\title{hts_schema}
\format{Returns a list which defines the schema for the HTS Disagg tool

HTS Schemas of the Data Pack Excel sheets}
\usage{
data(hts_schema)
}
\description{
Schema of the HTS template
}
\keyword{datasets}
