% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impatt}
\alias{impatt}
\title{impatt}
\format{An object of class \code{list} of length 3.}
\usage{
impatt
}
\value{
Returns a three column data frame with code,name and Data Pack code (dp_code)
}
\description{
List of option codes for PSNU prioritization
}
\keyword{datasets}
