% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack-importer.R
\name{ImportFollowOnMechs}
\alias{ImportFollowOnMechs}
\title{ImportFollowOnMechs(wb_info)}
\usage{
ImportFollowOnMechs(wb_info)
}
\arguments{
\item{wb_info}{Workbook info object.}
}
\value{
A data  frame with three columns Closing Out, Follow On, Notes.
If this sheet is blank, returns NULL.
}
\description{
Imports the follow on mechs sheet.
}
