% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack-importer.R
\name{ImportSheet}
\alias{ImportSheet}
\title{ImportSheet(wb_path,sheet_name,wb_type)}
\usage{
ImportSheet(wb_path, sheet_name, wb_type = "NORMAL")
}
\arguments{
\item{wb_path}{The absolute file path to the workbook.}

\item{sheet_name}{Name of the sheet to import.}

\item{wb_type}{Should be either NORMAL or HTS.}
}
\value{
Returns a data frame with the following columns:
#' \itemize{psnuuid,mechid,type,variable,value
  \item psnuid: Name of the PSNU or cluster
  \item mechid: Mechanism number
}
}
\description{
Imports a single sheet from a workbook.
}
