% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSheet.R
\name{ImportSheets}
\alias{ImportSheets}
\title{ImportSheets(wb_path,wb_type)}
\usage{
ImportSheets(wb_path, wb_type = "NORMAL")
}
\arguments{
\item{wb_path}{The absolute file path to the workbook.}

\item{wb_type}{Should be either NORMAL or HTS.}
}
\value{
Returns a boolean value TRUE if the the workbook is valid, otherwise false
}
\description{
Imports a single sheet from a workbook.
}
