% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack-importer.R
\name{ImportSheets}
\alias{ImportSheets}
\title{ImportSheets(wb_path)}
\usage{
ImportSheets(wb_path)
}
\arguments{
\item{wb_path}{The absolute file path to the workbook.}
}
\value{
Returns a list of data frames: 
#'  \itemize{
           \item wb_info: Workbook Info
           \item data: Standard d2importer data frame
           \item follow_on_mechs: Data frame of follow on mechs.
           }
}
\description{
Imports all sheets from the workbook
}
