% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main_schema}
\alias{main_schema}
\title{main_schema}
\format{An object of class \code{list} of length 2.}
\usage{
data(main_schema)
}
\value{
Returns a list which defines the schema for the  Disagg tool
}
\description{
Schema of the normal template
}
\keyword{datasets}
