% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rCOP18deMapT}
\alias{rCOP18deMapT}
\title{rCOP18deMapT}
\format{An object of class \code{data.frame} with 5275 rows and 26 columns.}
\usage{
rCOP18deMapT
}
\value{
Returns an object with the following columns
}
\description{
Object used for further processing and matching of data pack codes
}
\keyword{datasets}
