% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack-importer.R
\name{ValidateImpattSheet}
\alias{ValidateImpattSheet}
\title{ValidateImpattSheet(d,wb_info)}
\usage{
ValidateImpattSheet(d, wb_info)
}
\arguments{
\item{d}{A parsed data frame with IMPATT data}
}
\description{
Validates the impatt sheet for completeness.
}
