% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack-importer.R
\name{ValidateSheet}
\alias{ValidateSheet}
\title{ValidateSheet(wb_path,schema)}
\usage{
ValidateSheet(wb_path, schema)
}
\arguments{
\item{wb_path}{The absolute file path to the workbook.}

\item{schema}{Specific schema for the sheet}
}
\value{
Returns a boolean value TRUE if the sheet is valid, otherwise, FALSE.
}
\description{
Validates the layout of a single sheet based on its schema definition.
}
