% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSheet.R
\name{ValidateSheet}
\alias{ValidateSheet}
\title{ValidateSheet(wb_path,sheet_name,wb_type)}
\usage{
ValidateSheet(wb_path, sheet_name, wb_type = "NORMAL")
}
\arguments{
\item{wb_path}{The absolute file path to the workbook.}

\item{sheet_name}{The name of the specific sheet to be validated within the workbook.}

\item{wb_type}{Should be either NORMAL or HTS}
}
\value{
Returns a boolean value TRUE if the sheet is valid, otherwise, FALSE.
}
\description{
Validates the layout of a single sheet based on its schema definition.
}
