% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack-importer.R
\name{ValidateSheets}
\alias{ValidateSheets}
\title{ValidateSheets(schemas,sheets)}
\usage{
ValidateSheets(schemas, sheets, wb_info)
}
\arguments{
\item{schemas}{Schemas for this workbook}

\item{sheets}{Names of sheets}

\item{wb_info}{Workbook info for the worbook.}
}
\value{
Returns a boolean value TRUE if the sheet is valid, otherwise, FALSE.
}
\description{
Validates all of the sheets
}
