% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pack-importer.R
\name{ValidateWorkbook}
\alias{ValidateWorkbook}
\title{ValidateWorkbook(wb_path)}
\usage{
ValidateWorkbook(wb_path, distribution_method = NA,
  support_files_path = NA)
}
\arguments{
\item{wb_path}{The absolute file path to the workbook.}

\item{distribution_method}{The distribution method to use.}

\item{support_files_path}{Path to the support files directory.}
}
\value{
Returns an object with information about the workbook, if the file is valid.
Otherwise, the function will produce an error.
}
\description{
Validates the layout of all relevant sheets in a data pack workbook
}
