% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_site_level_tool.R
\name{write_site_level_sheet}
\alias{write_site_level_sheet}
\title{write_site_level_sheet(wb,schema,df)}
\usage{
write_site_level_sheet(wb, schema, d)
}
\arguments{
\item{wb}{Workbook to be written to.}

\item{schema}{Schema object for this sheet.}

\item{d}{Data frame object.}
\item{wb}{Workbook object}

\item{schema}{Schema object for this sheet}

\item{d}{Data frame object}
}
\description{
Validates the layout of all relevant sheets in a data pack workbook
}
